Refactored repositories and improved MemberGate.kt

+++ src/main/kotlin/be/duncanc/discordmodbot/bot/services/MemberGate.kt
@@ -20,8 +20,8 @@ import be.duncanc.discordmodbot.bot.commands.CommandModule
 import be.duncanc.discordmodbot.bot.sequences.Sequence
 import be.duncanc.discordmodbot.bot.utils.limitLessBulkDelete
 import be.duncanc.discordmodbot.data.entities.GuildMemberGate
-import be.duncanc.discordmodbot.data.entities.MemberGateQuestion
-import be.duncanc.discordmodbot.data.repositories.MemberGateQuestionRepository
+import be.duncanc.discordmodbot.data.redis.hash.MemberGateQuestion
+import be.duncanc.discordmodbot.data.repositories.key.value.MemberGateQuestionRepository
 import be.duncanc.discordmodbot.data.services.MemberGateService
 import net.dv8tion.jda.api.MessageBuilder
 import net.dv8tion.jda.api.Permission
@@ -42,8 +42,8 @@ import kotlin.collections.ArrayList
  * Welcomes users when the join, get accepted and makes them answer questions before they get accepted.
  */
 @Component
-class MemberGate
-internal constructor(
+@Transactional
+class MemberGate(
         private val memberGateQuestionRepository: MemberGateQuestionRepository,
         private val memberGateService: MemberGateService
 ) : CommandModule(
@@ -101,14 +101,24 @@ internal constructor(
         val gateTextChannel = memberGateService.getGateChannel(event.guild.idLong, event.jda)
         val welcomeChannel = memberGateService.getWelcomeChannel(event.guild.idLong, event.jda)
         if (gateTextChannel != null) {
-            gateTextChannel.sendMessage(
-                    "Welcome " + event.member.asMention + ", this server requires you to read the " +
-                            (memberGateService.getRulesChannel(event.guild.idLong, event.jda)?.asMention
-                                    ?: "rules") +
-                            " and answer a question regarding those before you gain full access.\n\n" +
-                            "If you have read the rules and are ready to answer the question, type ``!" + super.aliases[1] + "`` and follow the instructions from the bot.\n\n" +
-                            "Please read the pinned message for more information."
-            ).queue { message -> message.delete().queueAfter(5, TimeUnit.MINUTES) }
+            val memberGateQuestion = memberGateQuestionRepository.findById(event.user.idLong).orElse(null)
+            if (memberGateQuestion != null) {
+                gateTextChannel.sendMessage(
+                        """
+                        Welcome back ${event.member.asMention}. We stored your last answer for you.
+                        A moderator can review it using ```!review ${event.member.id}
+                        """.trimIndent()
+                )
+            } else {
+                gateTextChannel.sendMessage(
+                        "Welcome " + event.member.asMention + ", this server requires you to read the " +
+                                (memberGateService.getRulesChannel(event.guild.idLong, event.jda)?.asMention
+                                        ?: "rules") +
+                                " and answer a question regarding those before you gain full access.\n\n" +
+                                "If you have read the rules and are ready to answer the question, type ``!" + super.aliases[1] + "`` and follow the instructions from the bot.\n\n" +
+                                "Please read the pinned message for more information."
+                ).queue { message -> message.delete().queueAfter(5, TimeUnit.MINUTES) }
+            }
         } else if (welcomeChannel != null) {
             val welcomeMessages = memberGateService.getWelcomeMessages(event.guild.idLong).toTypedArray()
             if (welcomeMessages.isNotEmpty()) {
@@ -177,7 +187,7 @@ internal constructor(
         }
 
         if (memberGateQuestionRepository.existsById(member.idLong)) {
-            event.channel.sendMessage("You have already tried answering a question. A moderator now needs to manually review you. Please state that you have read and agree to the rules and need manual approval.")
+            event.channel.sendMessage("You have already tried answering a question. A moderator now needs to manually review you. Please be patient.")
                     .queue { it.delete().queueAfter(1, TimeUnit.MINUTES) }
             return
         }
@@ -573,7 +583,7 @@ internal constructor(
         init {
             memberGateQuestionRepository.findById(userId).ifPresentOrElse({
                 val userQuestionAndAnswer = it.question + '\n' + it.answer
-                if (userQuestionAndAnswer != null) {
+                if (userQuestionAndAnswer.isNotBlank()) {
                     val message: Message =
                             MessageBuilder().append("The user answered with the following question:\n")
                                     .appendCodeBlock(userQuestionAndAnswer, "text")
@@ -598,20 +608,20 @@ internal constructor(
             if (!memberGateQuestionRepository.existsById(userId)) {
                 throw IllegalStateException("The user is no longer in the queue; another moderator may have reviewed it already.")
             }
-                when (val messageContent: String = event.message.contentDisplay.toLowerCase()) {
-                    "approve", "accept" -> {
-                        accept(event)
-                    }
-                    "reject", "refuse" -> {
-                        reject(event)
-                    }
-                    "noop" -> {
-                        reject(event, noOp = true)
-                    }
-                    else -> {
-                        throw IllegalArgumentException("Expecting one of the previously mentioned responses, but got \"$messageContent\" as response")
-                    }
+            when (val messageContent: String = event.message.contentDisplay.toLowerCase()) {
+                "approve", "accept" -> {
+                    accept(event)
                 }
+                "reject", "refuse" -> {
+                    reject(event)
+                }
+                "noop" -> {
+                    reject(event, noOp = true)
+                }
+                else -> {
+                    throw IllegalArgumentException("Expecting one of the previously mentioned responses, but got \"$messageContent\" as response")
+                }
+            }
         }
 
         private fun reject(event: MessageReceivedEvent, noOp: Boolean = false) {
